Calculate normalized counts.

Calculate normalized counts.

normalize_counts(x, ...)

# S4 method for BbcSE
normalize_counts(x, de_pkg = "edger")

# S4 method for DGEList
normalize_counts(x)

Arguments

x
A BbcSE object or a DGEList or a DESeqDataSet.

...
Not used currently

de_pkg
"edger" or "deseq2". Only used if x is a BbcSE

Value

A BbcSE object or a matrix or....

Details

Normalization uses methods from the package corresponding to the de_pkg paramter:

edger

Uses edgeR::cpm with normalized.lib.sizes = TRUE and log = TRUE

deseq2

Not implemented yet.

Methods (by class)

BbcSE: For de_pkg="edger", a SummarizedExperiment object is created and stored in the BbcEdgeR object in the edger slot. Group average log normalized counts are also calculated and stored as rowData.

DGEList: log normalized counts returned as a matrix.

See also