Plot heatmap

Plot heatmap with option for splitting rows and columns and adding annotation.

plot_heatmap(
  x,
  genes = rownames(x),
  de_method = "edger",
  assay_name = "norm_log_cpm",
  gene_labels = NULL,
  coldata_annot = NULL,
  coldata_split = NULL,
  rowdata_annot = NULL,
  rowdata_split = NULL,
  clust_rows = TRUE,
  clust_cols = TRUE,
  clustering_distance_rows = "euclidean",
  clustering_distance_columns = "euclidean",
  zscores = FALSE,
  grouped = FALSE
)

Arguments

x
A BbcSE object

genes
gene IDs (rownames)

de_method
"edger" or "deseq2"

assay_name
Name of assay from the 'norm_cts' slot. Not compatible with grouped=TRUE. Option was implemented with the idea of being able to plot batch-corrected data.

gene_labels
a column name from rowData. Must contain unique values

coldata_annot
character vector of colData colnames for annotating

coldata_split
character value of colData colname for splitting

rowdata_annot
character vector of rowData colnames for annotating

rowdata_split
character value of rowData colname for splitting

clust_rows
logical for whether rows should be clustered

clust_cols
logical for whether columns should be clustered

clustering_distance_rows
see 'clustering_distance_rows' in ComplexHeatmap::Heatmap()

clustering_distance_columns
see 'clustering_distance_columns' in ComplexHeatmap::Heatmap()

zscores
logical indicating whether the expression matrix should be converted to Z-scores

grouped
logical indicating whether the mean expression per group should be shown.

Value

A Heatmap-class object.

See also